backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai>=0.8.2
pymupdf==1.24.10
Pillow==10.4.0
